requirements.txt
streamlit==1.27.2
tensorflow==2.10.1
mtcnn==0.1.1
opencv-python-headless==4.8.1.78
numpy==1.26.4
protobuf==3.19.6
Pillow==10.0.0
altair==4.2.2
